+++ cuttlepool/__init__.py
@@ -2,7 +2,7 @@
 """
 Cuttle Pool, a SQL pool implementation.
 
-:license: MIT, see LICENSE for details.
+:license: BSD 3-clause, see LICENSE for details.
 """