+++ CHANGELOG.rst
@@ -4,6 +4,11 @@ Cuttle Pool Changelog
 
 Here are the changes made to Cuttle Pool for each release.
 
+Version 0.6.0
+-------------
+
+Minor release, unreleased
+
 Version 0.5.0
 -------------